backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
cryptography==41.0.5
